Players get a win/loose announcment.
Revised Prefabs for the fleet menu.
Added a scene changer script, that loads all necessary scenes in the game.
Added a script, that announces the players, if they have won or lost the game.
On meeting a win condition, a scene that resolves the game is loaded.
Refactured the fleet menu script.